commited by kimlang

+++ resources/views/cp/promotion/index.blade.php
@@ -59,6 +59,25 @@
 			        }
 				});
 			}
+			function showIsDelete(id){
+			active = 0;
+			$.ajax({
+			        url: "{{ route($route.'.update-delete-status') }}",
+			        method: 'POST',
+			        data: {id:id, active:active },
+			        success: function( response ) {
+			            if ( response.status === 'success' ) {
+			            	swal("Nice!", response.msg ,"success");
+			            	location.reload();
+			            }else{
+			            	swal("Error!", "Sorry there is an error happens. " ,"error");
+			            }
+			        },
+			        error: function( response ) {
+			           swal("Error!", "Sorry there is an error happens. " ,"error");
+			        }
+				});
+		}
 	</script>	
 @endsection
 
@@ -105,7 +124,16 @@
 				<th>Title</th>
 				<th>Published</th>
 				<th>Image</th>
+				@if(Auth::user()->position_id == 1)
+				<th>Created By</th>
+				<th>Created By</th>
+				<th>Updated By</th>
+				@endif
 				<th>Updated Date</th>
+				@if(Auth::user()->position_id == 1)
+						<th>Delete By</th>
+						<th>Delete Date</th>
+						@endif
 				<th></th>
 			</tr>
 		</thead>
@@ -130,12 +158,28 @@
 							No Image Avaiable
 						@endif
 					</td>
+					@if(Auth::user()->position_id == 1)
+							<td>{{ $row->creator->name }}</td>
+							<td>{{ $row->created_at }}</td>
+							<td>@if($row->updater != '') {{ $row->updater->name }} @else no updater @endif</td>
+							@endif
 					<td>{{ $row->updated_at }}</td>
+					@if(Auth::user()->position_id == 1)
+								<td>@if($row->deleter != '') {{ $row->deleter->name }} @else no deleter @endif</td>
+								<td>{{ $row->deleted_at }}</td>
+							@endif
 					<td style="white-space: nowrap; width: 1%;">
 						<div class="tabledit-toolbar btn-toolbar" style="text-align: left;">
                            	<div class="btn-group btn-group-sm" style="float: none;">
                            		<a href="{{ route($route.'.edit', $row->id) }}" class="tabledit-edit-button btn btn-sm btn-success" style="float: none;"><span class="fa fa-eye"></span></a>
-                           		<a href="#" onclick="deleteConfirm('{{ route($route.'.trash', $row->id) }}', '{{ route($route.'.index') }}')" class="tabledit-delete-button btn btn-sm btn-danger" style="float: none;"><span class="glyphicon glyphicon-trash"></span></a>
+                           		@if($row->is_deleted == 0)
+		                           		<a href="#" onclick="deleteConfirm('{{ route($route.'.trash', $row->id) }}', '{{ route($route.'.index') }}')" class="tabledit-delete-button btn btn-sm btn-danger" style="float: none;"><span class="glyphicon glyphicon-trash"></span></a>
+		                           		@endif
+		                           		@if(Auth::user()->position_id == 1)
+		                           			@if($row->is_deleted == 1)
+ 											<button onclick="showIsDelete({{$row->id}})" class="tabledit-edit-button btn btn-sm btn-warning" style="float: none;"> <span class="fa fa-check"></span> Show</button>
+ 											@endif
+		                           		@endif
                            	</div>
                        </div>
                     </td>
